+++ bench/object/index.js
@@ -1,5 +1,5 @@
 'use strict';
 
 module.exports = {
-	assign: process.env.EXPERIMENTAL ? require('./assign') : undefined,
+	assign: require('./assign'),
 };